chore: Move form to component
